refactor(ast): move SrcInfo and related things to grammar.ts
